Adding forceMousedown method.
We need in our scene to have a mousedown event that happens prior to setting up this control and want the scene's mousedown event to do its work while at the same time letting this control begin its work. Therefore we need a method that forces this control's mousedown method to fire.